Inline ResultConsumer as nested class in Main.cs
Move ResultConsumer into the containing class (SceneDelegate or
AppDelegate) in each template, matching how Android's template
structures TestInstrumentation.cs. This eliminates the separate
ResultConsumer.cs file.
